# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CFP Energy - User Management Api",
        "contact": {
            "email": "rener.gbrl.p@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/",
            "description": "Local Url"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of projects",
                "description": "Returns list of projects",
                "operationId": "getusersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "user id",
                        "format": "string",
                        "default": "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
                    },
                    "first_name": {
                        "title": "first name",
                        "description": "first name",
                        "format": "string",
                        "default": "John"
                    },
                    "last_name": {
                        "title": "last name",
                        "description": "last name",
                        "format": "string",
                        "default": "Doe"
                    },
                    "username": {
                        "title": "username",
                        "description": "username",
                        "format": "string",
                        "default": "johndoe"
                    },
                    "phone_number": {
                        "title": "phone_number",
                        "description": "phone number",
                        "format": "string",
                        "default": "+5511999999999"
                    },
                    "date_of_birth": {
                        "title": "date_of_birth",
                        "description": "date of birth",
                        "format": "string",
                        "default": "yyyy-mm-dd"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "format": "string",
                        "default": "john.doe@email.com"
                    },
                    "profile_picture_url": {
                        "title": "profile_picture_url",
                        "description": "user's profile picture url",
                        "format": "string",
                        "default": "https://ui-avatars.com/api/?name=John+Doe&background=random&format=png"
                    },
                    "is_active": {
                        "title": "is_active",
                        "description": "status of user",
                        "format": "string",
                        "default": "true"
                    },
                    "roles": {
                        "title": "roles",
                        "description": "roles",
                        "format": "array",
                        "default": [
                            "user"
                        ]
                    },
                    "permissions": {
                        "title": "permissions",
                        "description": "permissions",
                        "format": "array",
                        "default": [
                            "view-users"
                        ]
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "date of creation of the user",
                        "format": "date",
                        "default": "yyyy-mm-dd"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "Credentials Based": {
                "type": "apiKey",
                "description": "Use your email / password combo to obtain a token (e.g.: Bearer {{token}})",
                "name": "Token",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/auth/login",
                        "refreshUrl": "/auth/refresh",
                        "scopes": {}
                    }
                },
                "scheme": "http"
            },
            "Cookie": {
                "type": "http",
                "description": "Cookie Based Authentication",
                "name": "Cookie",
                "in": "cookie",
                "flows": {
                    "password": {
                        "authorizationUrl": "/auth/login",
                        "refreshUrl": "/auth/refresh",
                        "scopes": {}
                    }
                },
                "scheme": "http"
            }
        }
    },
...
}